% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\docType{data}
\name{CryptoScale}
\alias{CryptoScale}
\title{Default coloring scale for all things CRYPTOCOCCUS}
\format{
An object of class \code{ScaleDiscrete} (inherits from \code{Scale}, \code{ggproto}, \code{gg}) of length 16.
}
\usage{
CryptoScale
}
\description{
Default coloring scale for all things CRYPTOCOCCUS
}
\keyword{datasets}
