% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{.get_offspring}
\alias{.get_offspring}
\title{Finds All DesendentsOf a Node}
\usage{
.get_offspring(phylotable, ancestor, env)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo that is being queried}

\item{ancestor}{The node you want to investigate}

\item{env}{The environment to which to write the answers}
}
\value{
A boolean set to TRUE if the node has no further descendents
}
\description{
This function takes in a phylo object or a tibble from a phylo object, an environment
to populate, the expectation being that it is empty, and a node to query. It returns a
NULL and modifies the environment that was passed. If no environment is passed, it assumes
that there is an environment called answer_env. Because this function could not initialize
and return this environment, there is a wrapper funtion that is exposed under the same name
without the starting dot. The environment is set to contain a range of variables with the
name of the node and a boolean assignment indicating if the node is a leaf. When testing,
uncomment the showme() calls in the code bellow and use capture.output(type="messages").
}
