% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{.printClade}
\alias{.printClade}
\title{Prints nwk format recursively}
\usage{
.printClade(phylotable, parent, depth = -1)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo that is being queried}

\item{ancestor}{The node you want to investigate}

\item{env}{The environment to which to write the answers}
}
\value{
A boolean set to TRUE if the node has no further descendents
}
\description{
depth is the key! -1 to print everything, 0 prints nothing
}
