% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{find_LCA}
\alias{find_LCA}
\title{Returns The Last Common Ancestor of a Set of Tips}
\usage{
find_LCA(phylotable, tiplist = NULL, nodelist = NULL)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo. fmt as: parent,node,\<labe\>}

\item{tiplist}{Char vectorwith node lables to include.}

\item{nodelist}{Int vector with nodes to include.}
}
\value{
Int representing last common ancestor
}
\description{
At least one of the
}
