% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{find_root}
\alias{find_root}
\title{Get the root of a phylogeny}
\usage{
find_root(phylotable)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo that is being queried}
}
\value{
An integer representing the root node of a phylogeny
}
\description{
This function takes in a phylo object or a tibble from a phylo object and
it returns the integer of the node with no further parents
}
