% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_children}
\alias{get_children}
\title{Get Children Nodes}
\usage{
get_children(phylotable, parent)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo  that is being queried}

\item{parent}{The node you wish to find the children for}
}
\value{
An integer vector corresponding to the children nodes
}
\description{
This function takes in a phylo object or a tibble from a phylo object and
returns an integer vector with the direct children of the node.
}
