% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_ChildTipLabels}
\alias{get_ChildTipLabels}
\title{Returns all tips for a set of node}
\usage{
get_ChildTipLabels(phylotable, nodelist)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo. fmt as: parent,node,\<labe\>}

\item{nodelist}{Int vector with nodes to delete.}
}
\value{
Imodified phylotable
}
\description{
useful to use droptips as remove_node seems to create problems
}
