% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_offspring}
\alias{get_offspring}
\title{Get All Descendants of a Node}
\usage{
get_offspring(phylotable, ancestor)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo that is being queried}

\item{ancestor}{The node crowning the clade you seek}
}
\value{
An integer vector with all the nodes in the clade. This vector has the attribute "leaf" which describes with a boolean the status of each vector as a tip.
}
\description{
This function takes in a phylo object or a tibble from a phylo object and
it returns the integer of the node with no further parents
}
