% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_parent}
\alias{get_parent}
\title{Get Parent Node}
\usage{
get_parent(phylotable, node)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo  that is being queried}

\item{node}{The node you wish to find the parent for}
}
\value{
An integer corresponding to the parent node
}
\description{
This function takes in a phylo object or a tibble from a phylo object and
returns the parent of the node.
}
