% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_siblings}
\alias{get_siblings}
\title{Get sibling Nodes}
\usage{
get_siblings(phylotable, node)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo  that is being queried}

\item{node}{The node you wish to find the siblings for}
}
\value{
An integer vector corresponding to the sibling nodes; empty if the clade has no siblings
}
\description{
This function takes in a phylo object or a tibble from a phylo object and
returns an integer vector with the direct siblings of the node.
}
