% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{is_leaf}
\alias{is_leaf}
\title{Checks If Node Is Leaf}
\usage{
is_leaf(phylotable, query)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo that is being queried}

\item{query}{The node you want to investigate}
}
\value{
A boolean set to TRUE if the node has no further descendents
}
\description{
This function takes in a phylo object or a tibble from a phylo object and
a node to query. It returns a boolean indicating if the node is a leaf.
}
