% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{remove_node}
\alias{remove_node}
\title{deletes nodes and descendants.}
\usage{
remove_node(phylotable, nodelist)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo. fmt as: parent,node,\<labe\>}

\item{nodelist}{Int vector with nodes to delete.}
}
\value{
Imodified phylotable
}
\description{
careful not to use labels
}
