% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{replaceNode}
\alias{replaceNode}
\title{Makes the necessary changes to a phylotable to customize nodes}
\usage{
replaceNode(phylotable, old, replacement, .checkForConflicts = TRUE)
}
\arguments{
\item{phylotable}{Phylo object or tibble from phylo. fmt as: parent,node,\<labe\>}

\item{old}{Int representing old node number}

\item{replacement}{Int value to place in}

\item{checkForConflicts}{Bool value to limit the recursive calls. Do not use. this function operates under the expectation that once the nodes are replaced with random numbers, there should not be more conflicts.}
}
\value{
a  phylo tibble with the requested changes
}
\description{
Makes the necessary changes to a phylotable to customize nodes
}
